LayoutTests/imported/w3c/web-platform-tests/css/cssom/stylesheet-title-expected.txt

Should be green


PASS Preferred style sheet name
FAIL StyleSheet.title assert_equals: expected 4 but got 3
